Update errors/incomplete regexes
While the other formatter parsing options looked for both cases, Errors and Incomplete were
left as only capitalized. As of PHPUnit 4.8 (used for testing this) the Errors and Incomplete
messages are all lower cases, so the formatter was not picking them up and was providing
incorrect information to the terminal-notifier.  This fix adds lower cased versions to the
existing regexes to match the rest of the parsing options.